chore: move CreateDestination interop behaviour behind interop flag
