+++ pkg/didcomm/common/service/destination.go
@@ -8,13 +8,8 @@ package service
 
 import (
 	"fmt"
-	"strings"
 
-	"github.com/btcsuite/btcutil/base58"
-
-	diddoc "github.com/hyperledger/aries-framework-go/pkg/doc/did"
 	vdrapi "github.com/hyperledger/aries-framework-go/pkg/framework/aries/api/vdr"
-	"github.com/hyperledger/aries-framework-go/pkg/vdr/fingerprint"
 )
 
 // Destination provides the recipientKeys, routingKeys, and serviceEndpoint for an outbound message.
@@ -28,8 +23,6 @@ type Destination struct {
 
 const (
 	didCommServiceType = "did-communication"
-	// legacyDIDCommServiceType is the non-spec service type used by legacy didcomm agent systems.
-	legacyDIDCommServiceType = "IndyAgent"
 )
 
 // GetDestination constructs a Destination struct based on the given DID and parameters
@@ -41,68 +34,4 @@ func GetDestination(did string, vdr vdrapi.Registry) (*Destination, error) {
 	}
 
 	return CreateDestination(docResolution.DIDDocument)
-}
-
-// CreateDestination makes a DIDComm Destination object from a DID Doc as per the DIDComm service conventions:
-// https://github.com/hyperledger/aries-rfcs/blob/master/features/0067-didcomm-diddoc-conventions/README.md.
-func CreateDestination(didDoc *diddoc.Doc) (*Destination, error) {
-	didCommService, ok := diddoc.LookupService(didDoc, didCommServiceType)
-	if !ok {
-		// Interop: fallback to using IndyAgent service type
-		didCommService, ok = diddoc.LookupService(didDoc, legacyDIDCommServiceType)
-		if !ok {
-			return nil, fmt.Errorf("create destination: missing DID doc service")
-		}
-	}
-
-	if didCommService.ServiceEndpoint == "" {
-		return nil, fmt.Errorf("create destination: no service endpoint on didcomm service block in diddoc: %+v", didDoc)
-	}
-
-	if len(didCommService.RecipientKeys) == 0 {
-		return nil, fmt.Errorf("create destination: no recipient keys on didcomm service block in diddoc: %+v", didDoc)
-	}
-
-	// Interop: service keys that are raw base58 public keys should be converted to did:key format
-	return &Destination{
-		RecipientKeys:     convertAnyB58Keys(didCommService.RecipientKeys),
-		ServiceEndpoint:   didCommService.ServiceEndpoint,
-		RoutingKeys:       convertAnyB58Keys(didCommService.RoutingKeys),
-		MediaTypeProfiles: didCommService.Accept,
-	}, nil
-}
-
-func convertAnyB58Keys(keys []string) []string {
-	var didKeys []string
-
-	for _, key := range keys {
-		if key == "" {
-			didKeys = append(didKeys, key)
-			continue
-		}
-
-		// skip if the key is a relative did-url (ie, it starts with ?, /, or #)
-		if strings.Contains("?/#", string(key[0])) { // nolint:gocritic
-			didKeys = append(didKeys, key)
-			continue
-		}
-
-		// skip if the key is already a did
-		if strings.HasPrefix(key, "did:") {
-			didKeys = append(didKeys, key)
-			continue
-		}
-
-		rawKey := base58.Decode(key)
-		if len(rawKey) == 0 {
-			didKeys = append(didKeys, key)
-			continue
-		}
-
-		didKey, _ := fingerprint.CreateDIDKey(rawKey)
-
-		didKeys = append(didKeys, didKey)
-	}
-
-	return didKeys
 }
